+++ src/asserts.h
@@ -17,6 +17,9 @@ void assertArgIsString(VM *vm, const char *function, Value *args, int index);
 void assertNonZero(VM *vm, const char *function, double number, int index);
 void assertPositiveNumber(VM *vm, const char *function, double number,
                           int index);
+void assertNonNegativeNumber(VM *vm, const char *function, double number,
+                             int index);
+void assertWholeNumber(VM *vm, const char *function, double number, int index);
 void raiseError(VM *vm, const char *message);
 
 #endif